requirements.txt
llama-index==0.11.7
qdrant-client==1.11.1
llama-index-embeddings-fastembed==0.2.0
llama-index-vector-stores-qdrant==0.3.0
streamlit==1.38.0
scikit-learn==1.5.2
sentence_transformers==3.1.1
python-dotenv==1.0.1
seaborn==0.13.2
matplotlib==3.9.2
